pyproject.toml
[project]
name = "webl234s5"
version = "0.1.0"
authors = [{ name = "w1vern", email = "bndvvl@gmail.com" }]
requires-python = ">=3.13"

dependencies = [
    "accelerate>=1.10.1",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.117.1",
    "fastapi-controllers>=0.3.0",
    "faststream[rabbit]>=0.5.48",
    "httpx>=0.28.1",
    "openai>=1.109.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "uvicorn[standard]>=0.37.0",
    "websockets>=15.0.1",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
]


